fix poetry build not including version_sets/0.5 files
Seems like it has to do with the fact that they are symlinks.  By adding a
`build.py` which puts them into places based on the contents of
`version_sets.yaml`, we can get the benefit of storing all rules in
base_rule_definitions and pulling them in to various versions of the FOCUS
spec.

Signed-off-by: Joshua Kwan <joshk@ternary.app>